@@ -19,12 +19,12 @@
   ],
   "require": {
     "php": ">=5.5",
-    "laravel/framework": "5.1.*|5.2.*|5.3.*|5.4.*|5.5.*"
+    "laravel/framework": "5.5.*"
   },
   "require-dev": {
-    "phpunit/phpunit": "~5.7",
+    "phpunit/phpunit": "~6.0",
     "mockery/mockery": "~0.9",
-    "orchestra/testbench": "^3.1|^3.2|^3.3|^3.4",
+    "orchestra/testbench": "^3.5",
     "phpro/grumphp": "^0.11",
     "friendsofphp/php-cs-fixer": "^2.4"
   },
