@@ -4,7 +4,7 @@ namespace Nwidart\Modules\tests;
 
 use Nwidart\Modules\Support\Migrations\NameParser;
 
-class NameParserTest extends \PHPUnit_Framework_TestCase
+class NameParserTest extends \PHPUnit\Framework\TestCase
 {
     /** @test */
     public function it_gets_the_original_name()
