@@ -4,7 +4,7 @@ namespace Nwidart\Modules\tests;
 
 use Nwidart\Modules\Support\Migrations\SchemaParser;
 
-class SchemaParserTest extends \PHPUnit_Framework_TestCase
+class SchemaParserTest extends \PHPUnit\Framework\TestCase
 {
     /** @test */
     public function it_generates_migration_method_calls()
