@@ -24,7 +24,7 @@ class DisableCommand extends Command
     /**
      * Execute the console command.
      */
-    public function fire()
+    public function handle()
     {
         $module = $this->laravel['modules']->findOrFail($this->argument('module'));
 
