@@ -26,7 +26,7 @@ class MakeCommand extends Command
     /**
      * Execute the console command.
      */
-    public function fire()
+    public function handle()
     {
         $names = $this->argument('name');
 
