@@ -28,7 +28,7 @@ class MigrateRefreshCommand extends Command
     /**
      * Execute the console command.
      */
-    public function fire()
+    public function handle()
     {
         $this->call('module:migrate-reset', [
             'module' => $this->getModuleName(),
