@@ -34,7 +34,7 @@ class MigrateRollbackCommand extends Command
     /**
      * Execute the console command.
      */
-    public function fire()
+    public function handle()
     {
         $this->module = $this->laravel['modules'];
 
