@@ -23,7 +23,7 @@ class SetupCommand extends Command
     /**
      * Execute the console command.
      */
-    public function fire()
+    public function handle()
     {
         $this->generateModulesFolder();
 
@@ -35,7 +35,8 @@ class SetupCommand extends Command
      */
     public function generateModulesFolder()
     {
-        $this->generateDirectory($this->laravel['modules']->config('paths.modules'),
+        $this->generateDirectory(
+            $this->laravel['modules']->config('paths.modules'),
             'Modules directory created successfully',
             'Modules directory already exist'
         );
@@ -46,7 +47,8 @@ class SetupCommand extends Command
      */
     public function generateAssetsFolder()
     {
-        $this->generateDirectory($this->laravel['modules']->config('paths.assets'),
+        $this->generateDirectory(
+            $this->laravel['modules']->config('paths.assets'),
             'Assets directory created successfully',
             'Assets directory already exist'
         );
