@@ -27,7 +27,7 @@ class UpdateCommand extends Command
     /**
      * Execute the console command.
      */
-    public function fire()
+    public function handle()
     {
         $this->laravel['modules']->update($name = $this->getModuleName());
 
