@@ -34,7 +34,7 @@ class Updater extends Runner
             $concatenatedPackages .= "\"{$name}:{$version}\" ";
         }
 
-        if(!empty($concatenatedPackages)) {
+        if (!empty($concatenatedPackages)) {
             $this->run("composer require {$concatenatedPackages}");
         }
     }
@@ -51,7 +51,7 @@ class Updater extends Runner
             $concatenatedPackages .= "\"{$name}:{$version}\" ";
         }
 
-        if(!empty($concatenatedPackages)) {
+        if (!empty($concatenatedPackages)) {
             $this->run("composer require --dev {$concatenatedPackages}");
         }
     }
