@@ -99,7 +99,7 @@ class MyAwesomeCommand extends Command
      *
      * @return mixed
      */
-    public function fire()
+    public function handle()
     {
         //
     }
